PMI-503: Move utility modules into `performance` package. We can always move them back to a dedicated package in the future if needed.
